+++ lib/util.js
@@ -19,9 +19,7 @@ function getPostConfig(json){
 }
 
 function sendSuccess(node, msg, res){
-    if(res.data){
-        msg.payload = res.data;
-    }
+    msg.payload = res.data;
     node.send(msg);
     node.status({
         fill: "blue",
@@ -31,13 +29,27 @@ function sendSuccess(node, msg, res){
 }
 
 function sendError(node, err){
-    const payload = err.response.data.message;
+    const payload = err.response.data;
     node.error(payload);
     node.status({
         fill: "red",
         shape: "ring",
         text: payload
     });
+}
+
+function generateOpts(n){
+    let opts = {};
+    if (nodes.getNode(n.device)){
+        opts.address = nodes.getNode(n.device).address;
+    }
+    if(n.service){
+        opts.service = n.service;
+    }
+    if(n.characteristic){
+        opts.characteristic = n.characteristic;
+    }
+    return opts;
 }
 
 module.exports = {
@@ -47,5 +59,6 @@ module.exports = {
     },
     getPostConfig: getPostConfig,
     sendSuccess: sendSuccess,
-    sendError: sendError
+    sendError: sendError,
+    generateOpts: generateOpts
 }
